Change operator precedence to make "op_w" bind almost as tightly as "."
The idea is to make whitespace-conjoined macros more broadly useful,
as it is expected that many desirable cases will be closely-binding,
and in the case that you need a complex argument to such a macro it
could be parenthesized to be treated as a single term.

This has the effect of making parens required on many `if`/`while` exprs.