Remove an old loopy example that is no longer helpful.
It was useful while debugging early language progress, but now that
we have a nice in-language spec runner and can test that way, this
old example is just lying in wait to confuse newcomers with code
that takes a lot of space and complexity to not do anything useful
or purposeful. Kind of like that run on sentence, there ^

At any rate, good night baby example code! Thanks for the help!