Fix typechecking of runtime type checks/matching.
The previous typechecking was allowing some unsafe cases,
and was not allowing all safe cases properly.

The new approach intersects the right hand side type ignoring its cap(s),
then requires the intersection to be a subtype of the original.

This better mimics what is going on under the hood in codegen,
where we do the type check without any runtime knowledge of caps,
so we require the compile time check to be cap-safe after that intersection.

Resolves #57.