Add "unsafe recover" to pretend to support recover for supporting Pony code.
We don't support real, safe recover yet, since we don't have the system
we'll need to understand safety of the Verona kind using origin tracking,
so for now we punt and implement a magic "unsafe recover" that is generated
from the Pony grammar for the purposes of unblocking some Pony language testing.

Because this has not safety guarantees and is only for Pony language testing,
it should not be exposed in the Mare language yet.

This will allow us to highlight and prioritize what else is missing before
we can support the Pony language (e.g., tuples, Pony-style FFI, etc),
and we will eventually work our way back to working on recover for real
in a methodical, principled way rather than rushing through it.