Work in progress - add the AltInfer pass,
which will hopefully eventually replace the Infer pass.

This pass is an experimental refactoring of the Infer pass, tracking
the space of possibilities for a node's type as a Span, so that the Span
can be resolved quickly later for any given reification.
The expectation is that this approach will reduce compile times
and maybe also provide more powerful inference capabilities.

Currently this pass is not used for compilation, unless you use the
--pass=alt_infer flag when running the `mare` binary.