Remove "unreal" concept from Jumps pass analysis.
This was added by another contributor but I don't understand it.

There are no comments explaining what it is for, nor tests that
demonstrate what it does or fail without it being present.

Most importantly, this analysis is causing very poor compiler performance,
taking >3 seconds to run this pass on a simple Hello World program.

Once my work on the AltInfer pass is merged this will be the last major
contributor to poor compiler performance, so I am getting rid of it now.

If something like this is needed at some point in the future,
we can find a more performant way to implement the analysis,
driven by whatever the real use case was for this feature
(which again, is unknown to me at this time - I don't know what "unreal" means,
or what it is meant to imply for later analysis that depends on the Jumps pass).