Use new :serve_errors compile pass target for LSP server.
Previously this was using :completeness, which used to be the final
error-generating pass, but since the refactor it is now nowhere near final,
and the final error-generating passes do not depend on one another,
so there was no single target that encapsulated them all.

So we added the new :serve_errors pass target which does nothing of its own,
but it has dependencies which imply that all passes that can generate errors
relevant to an LSP server run must be executed.