Distinguish initial_cond and repeat_cond in internal Loop AST.
This will allow different kinds of loops in the future,
like the "repeat ... while ..." style of loop which runs
unconditionally on the first run.

And this also helps fix a corner case CodeGen bug when
using a yield expression inside of a `while` loop condition.
Distinguishing the two conditions makes sense because we
CodeGen them twice, in two different ways.