CodeGen extra alloca for each local in a continuation.
This makes it so that a debugger like lldb or gdb can easily
show the values when running a debug build. Prior to this,
yielding functions were quite difficult to debug.

Note that the extra alloca is a bit pesky in its redundancy,
but we expect LLVM to be able to optimize that redundancy away.