Update function/type declaration syntax to look more like the call site.
The initial whitespace-separated syntax felt more consistent with the
overall design of how declarations were "supposed" to work,
but it turns out to be just too unusual and arbitrary-feeling
that the function definition site requires whitespace between ident and params,
whereas the call site for the function requires that no whitespace be present.

Now they both require that no whitespace be present.