Avoid specifying details about valid capabilities at the parser level.
Knowledge of which capabilities exist in type expressions isn't really
the right level of knowledge for the parse grammar, because the grammar
is meant to know the minimum possible to build a correct parse tree.

This change will make the parser accept any identifier as valid
capability syntax, even if it turns out that later in the compiler
there is no semantic meaning known for that chosen identifier.

Knowledge of capabilities should come only in later parts of the compiler
as we start to understand what the syntax means.