Update Dockerfile to build libponyrt.bc faster & more simply.
Also removes the pony patches that are no longer needed,
leaving only one small patch in place to make it work on alpine.

Also updates part of the codegen for libponyrt, because one of the
runtime functions now has a new parameter that we need to pass in.
It's an optional optimization, but it's a mandatory C ABI change.
Here we add the parameter and leverage the associated optimization.

See https://github.com/ponylang/ponyc/pull/3653