Always use visit_pre in Macros pass - for consistency.
Unlike some other forms of visitation (which go depth-first),
it turns out that macro expansion is more amenable to breadth-first.
So we change it from using mixed `visit`/`visit_pre` to using all `visit_pre`
for consistency and predictability reasons.

This uncovered a bug in the copy-on-mutate visitation for Jump,
in which it was wrongly returning the original instance instead
of the newly mutated instance when its inner term had been changed via visitation.
This manifested as a bug in which macro expansion inside a return expression was broken.

So this commit also fixes that small bug and adds a test case to prevent regression.