Fix tracing of structs and turn two other classes into structs.
We hadn't implemented proper GC tracing of structs yet before this commit,
and converting these two classes into structs exposed that issue.

We trace a Savi struct value for GC like a Pony tuple is traced -
we trace each field of the struct in order.

The struct value itself doesn't need to be traced when it is stack-allocated.
It only needs to be traced when it is in a boxed wrapper,
such as when it is combined in a type union.