Add generic `Pair` type.
With one type argument, it uses the same type for both elements.
With two type arguments, it uses different types for the elements.

This type is intended to be used anywhere a pair of values is needed,
where there is no need for additional methods more specific to the type.