Refactor AST building so that bodies are not nested in AST::Declare.
This lets us better handle bodies in the declarator interpreter,
such as the case of `:fun` followed by `:yields` followed by a body
for the `:fun`, which was previously hacked together using logic
in the intrinsic for `:yields` in order to make it possible.

This is also necessary for indentation, because positioning the
body inside the AST::Declare tree when it is potentially interleaved
with other declarations wreaks havoc on indentation correction in
AST::Format.