Fix case of using a single-identifier macro as an argument.
This was previously carved out as a special case, to facilitate
some code that used `next` as a variable name in JsonParser,
but now I'm just changing the variable name because carving out
that special case was a wrong/misguided thing to do.

There may be more things to figure out with single-identifier macros
to make things more clean/consistent, but this minimal change at
least resolves the immediate issue blocking @tonchis.