Refactor the declarator interpreter to interpret as a separate step.
Now, we accept declarations into a tree until we finish a top-level one,
then we immediately descend into that declaration and interpret it
recursively, allowing us to ensure the full tree is there before doing so.

This will become important for custom declarators that need to reference
data/terms from nested declarations in the generated code for the outer
declaration.