Allow a :copies declaration in a :manifest to provide the copied name.
This allows, for example, a spec manifest which copies a lib manifest
to also be seen by other packages as the provider of that lib's name.

This can be used to add a spec dependency to a package that depends
on the package being tested. For example, the Random specs will
soon add a dependency on the DeterminsticRandom lib (which depends
in turn on the Random lib) in order to have deterministic random
test values for the derived values that the Random methods provide.