Detect missing transitive dependencies (and auto-fix with `--fix` flag).
This commit introduces the new `--fix` flag which can be used to
instruct the compiler to automatically fix any errors which have
"fix edits" attached to them, by applying those fix edits to the
relevant source files on disk.

Fix edits are surgical, and do not rewrite/format the entire file.

This commit also introduces a new kind of integration test case,
which can test for fixable errors, by first checking the
expected error messages, then running again with `--fix` and
expecting compilation to succeed, and expecting files to be
modified to match an expected "after" content.

See `spec/integration/README.md` for more details on how to
write such an integration test.

Note that this commit still does not test anything about the
transitive deps other than the fact that they are present.

Further checks can come in future work that leverages this
basic foundation.