Stop using libclang as a linking intermediary.
We want to craft the lib args directly, since libclang makes
a few too many harmful assumptions for our purposes.

For example, on MacOSX systems, libclang expects to try to find
the path to the `lld` linker on disk and complains if it is missing.
We are using an embedded `lld` linker, so we don't want to depend
on finding an external `lld` linker on disk.

The code in this commit is based on some examination of how libclang
works, with a lot of simplification, in part because we only support
a handful of platforms currently.