Refactor the order of passes, putting Populate before Macros.
This necessitated some changes to the Populate pass, including
making it depend on the Namespace pass instead of ReferType,
and also refactoring it to have better caching properties
so that it does not wreck caching of passes after it.

While refactoring caching properties, it was noticed that some
caching functions lack the standard printing when the `--print-perf`
CLI flag is used.