Fix a semantics issue with `:is` and `:copies`.
When copying into an abstract type (e.g. `:trait`), we should
include the copying of functions with no body, in order that
they be included in the requirements that the trait then
imposes on the final concrete type that needs to implement the trait.

This commit includes some cleanup in the `DeterministicRandom`
package, removing some "TODO" items related to workarounds
for this prior limitation in the function-copying semantics,
as implemented in the `Populate` pass.