Add initial `Bytes.Slice` type.
Currently this type does almost nothing, but in the future it will
expose most of the read-only methods present on the `Bytes` type.

This allows us to clean up the implementation of some things in
`ByteStream.ChunkedWriter` and `ByteStream.ChunkedSink`,
which will also pave the way for us to have a `ByteStream.Pipe`
in the future, by having the byte slices be stored in form that
is safe (pointer-tamper-proof) but also efficient and compatible
with `writev` system calls without copying into other data structures.

In a future change, when we add the missing methods from `Bytes`
into `Bytes.Slice`, we will also add corresponding `:trait`s for
all shared features, so that we can keep the two in sync and
reduce some of the code duplication of having two different ways
to deal with an immutable sequence of bytes.