Remove `Bytes.Slice` type, making it private in `IO` package.
On second thought, now that we were able to further clean up
the `ByteStream` package to no longer rely on this type in its
interface, we don't have a very compelling justification to
make it a public / core type right now.

If further justification is found, we could reverse this decision
and reinstate it as a public type at that time.