Add `ByteStream.Sink.ToReader`.
This class is an implementation of `ByteStream.Sink` which emits into
a given `ByteStream.Reader`, acting as the `ByteStream.Source` for it.
This is often useful for testing protocol reading and writing in test suites.

The `write_bytes` call writes to a local chunk buffer, which gets flushed
into the `ByteStream.Reader` when the `write_flush` method is called.

Note that this approach will incur costs for copying the chunks, because the
underlying data model of `ByteStream.Reader` is a single contiguous buffer.
Given that testing is the primary use case for synchronously connecting a
`ByteStream.Writer` to a `ByteStream.Reader`, this copying cost is
considered mostly unimportant.