Fetch deps from GitHub using git+HTTPS, not git+SSH.
It's not guaranteed that the machine we're using for `savi deps update`
has an SSH key configured with GitHub access, but HTTPS transport
should always work, and allows anonymous pulling of public repos.

This should have been using git+HTTPS from the start.
Using git+SSH was a mistake that I'm fixing now.