Add `--llvm-ir` flag to emit LLVM IR to a file.
This commit also removes the `--print-ir` flag which printed to stdout.

The LLVM IR is always so huge that it is inconvenient to work with
unless pasted into a file and opened in a text editor.

This change removes that middle manual step and emits directly to
a file whose name is the same as the emitted binary name, but with
a suffix of `.ll`.

This commit also changes the behavior of LLVM Module Verification bugs.
For the same reasons, it will write to a file instead of stdout.
It also now has a message asking the user to open an issue ticket
with the failure output as well as the emitted `.ll` file.