Remove deprecated `Platform` alias functions.
This commit also cleans up the `Platform.savi` file and adds
proper documentation comments for all functions/constants therein.

It also moves the remaining `TODO` comments in that file into the
CodeGen pass of the compiler instead, where they will be eventually
resolved when it's time to add Windows support.