Add `:ffi` function declarations.
This makes `:ffimodule` declarations obsolete - FFI functions
can now be declared on any kind of type, not just `:ffimodule`.

This commit also switches over all FFI declarations in the core library
to all be declared in a single `:module` named `_FFI`, which will
be the new convention for libraries that use FFI.

This commit also adds some additional privacy checks for FFI functions.
And it ensures that FFI functions will never be indirected by virtual
calls, allowing them to never show up in virtual tables, which helps
with cross-platform FFI functions not getting linked into the binary.

This commit also happens to fix a privacy check bug in that same area,
and fixes some code in `spec/core/BitArray.Spec.savi` that was
acidentally exxploiting the privacy check bug to reach a private
field of the BitArray` type in the tests for it.