Clean up and better encapsulate `AsioEvent`.
After this commit, the `AsioEvent` mechanisms no longer expose
an interface that can break safety.

We don't expose here in the core library any of the useful methods for
dealing with events, because those break memory safety. Any library that
wants to deal with creating, modifying, unsubscribing, or destroying events
will need to use FFI to access the runtime functions for doing so,
triggering the same manifest visibility as any other FFI usage
and encouraging users to audit those libraries to guarantee that
they expose a memory-safe interface and use capability security.

Also, we encapsulate the `flags` and `arg` parameters to `AsioEvent.Actor`
inside an `AsioEvent` struct, so that an intervening non-FFI actor
cannot modify the parameters to trick an unknowing FFI-using engine
to break safety just by changing the numbers. To prevent this, the
`AsioEvent` struct has a private constructor, and can thus only be
constructed by the inner ASIO mechanisms of the runtime.
The `flags` and `arg` field can be read, but not changed, so they
can be viewed but not tampered with.

All of this is in service of making it so that FFI-using
evented libraries can expose an API that is fully safe
and not possible to compromise from non-FFI code.