BREAKING CHANGE: `:member` names in an `:enum` get an automatic prefix.
Now an `:enum My.Example` with `:member Foo` will allow access to that
member via the name `My.Example.Foo` instead of just `Foo`.

That is, the name of the `:enum` type will be automatically used as
a prefix for the name of each `:member`.

---

Note that this change necessitates introducing a patched version
of the `Time` library into our codebase to allow our tests to pass,
because the real `Time` library has not yet been updated to match
these breaking changes.