Add support for `relative:../path/to/dep` dependencies.
This is primarily expected to be used for libraries that
have an integration test or example to be used with the
current version of the library in a nearby source directory
of the same project's source tree.

It could also be used in complex applications that are divided
into many in-source-tree libraries.