Add String interpolation syntax.
This commit adds `\(` as a special escape sequence in double-quoted
strings, which begins the interpolation of an arbitrary expression.
The `)` character marks the end of an open interpolation expression.

The final value inside the expression will be copied into the
built string using the `into_string` method, after first calculating
how much space to reserve using the `into_string_space` method.
Both of these methods must be implemented on the type for that value,
or else type-checking will fail.

Resolves #84.