Fix compiler alias analysis of the sendability of params with defaults.
Prior to this change, an `iso` param that has a default value
was wrongly being treated as having cap `iso'aliased` when
analyzing its sendability for safety, because the inference
system was treating the result of the assignment as being
the value (which is an alias for normal assignment).

After this fix, both `DEFAULTPARAM` and `<<=` are treated as
special forms of assignment whose result value is not an alias.