Refactor the `namespace` pass to track by package instead of file.
Now that we have the new package management system,
the set of types available in any given context is determined
by the dependencies declared for that package, rather than
being able to load different sets of dependencies for each source file.

We can now simplify and reduce the memory and CPU footprint of the
namespace pass as a result of that language change.