Go back to using the new LLVM PassBuilder for optimization.
This reverts PR #203, in which we temporarily disabled these
optimizations due to an LLVM bug. We no longer observe this bug
after upgrading LLVM from 13.0.0 to 14.0.3 (see previous commit).

We also move the emission of LLVM IR to ensure it happens after
optimizations (in the new place where they are, after linking to the
runtime bitcode and thus allowing link time optimization with it).