Another small fix for Windows linker pathing.
I made a mistake in the previous commit that I wasn't able to
notice and detect until after the release because I'm testing
this on a Windows CI machine using published Savi releases.

Hopefully this is the last thing I need to fix to get the CI
library-check builds working on Windows, but I wouldn't count on that.