Add `FloatingPoint.Format`, using the Ryū algorithm.
The [Ryū algorithm](https://dl.acm.org/doi/pdf/10.1145/3296979.3192369)
is a fast and accurate way to print floating-point numbers,
using some clever bit magic and lookup tables.

At this time we implement only the 64-bit version of the algorithm,
but in the future a 32-bit version could be implemented as well.
But for now, when printing `F32`s we will first convert to `F64`.