Mark as many LLVM functions as possible with private linkage.
We're not compiling a library, so we want to mark as many functions
as possible with private linkage, so LLVM can optimize them aggressively,
including inlining them, eliminating them if they are not used, and/or
marking them to use the fastest available calling convention.

Prior to this commit, the public runtime functions were all marked as
external, so they would miss out on optimizations if we don't do this.