Fix CI breakage related to `types_graph` pass.
This pass is not in use, and has a tricky-to-troubleshoot error
now happening on the `main` branch. Not sure how it got past PR CI,
but we need to do a quick fix to get the `main` branch working again.

This PR just disables part of the pass and comments out the tests
for the pass. We can fix these if we decide to bring this pass into use.