Fix a code generation bug for struct receivers.
Prior to this commit, when calling a method on a value that could
be either a struct or a class (both of which have that method),
then invalid LLVM IR could be generated.

This was causing failures in the HTTPServer example code in CI.

This commit fixes that bug and adds a regression test.