Fix breakage in `FloatingPoint.Format` from concurrent PRs.
PR #358 and and #365 were open concurrently, and after both
were merged there was a behavioral merge conflict (but not
a source-level merge conflict detectable by git).

`FloatingPoint.Format`, which was added in #365, depended on
code from `Integer.Format`, which was removed in #358.

This PR resolves the issue by copying the removed code into
`FloatingPoint.Format`.