Add `\U` 32-bit Unicode escape sequence for string literals.
Prior to this commit, we only had the `\u` 16-bit escape sequence.

These two escape sequence forms are taken from C conventions.

This commit also adds the `\'` escape sequence for `"` strings
and the `\"` escape sequence for `\'` strings, for consistency.
This allows the user to switch from one string type to the other
without needing to change their escape sequences, provided that
both forms of quotes are already escaped.