Remove `musl` target detection hack.
Still `llvm-config` prints the wrong host target for us,
But we now cover this need via the `LLVM_DEFAULT_TARGET` env var
that we set at compiler-compile time, which is based on our own
platform detection that we run as part of our `Makefile` setup.

This hack is actually detrimental now, because when it is in place,
it prevents cross-compiling from `-linux-gnu` to `-linux-musl`,
because the user-compile-time platform detection detects as `-gnu`
and overrides the explicit `-musl` requested by the user
when they supplied the `-X`/`--cross-compile` argument.