Deprecate `SDK_ROOT` env var, unifying with `SAVI_SYS_ROOT` env var.
Now all platforms support the ability to designate a sys root path
via the `SAVI_SYS_ROOT` env var. When this env var is specified,
then the linker step will only use search paths inside that directory.
If not specified, a default set of sys roots for that platform
will be used, which are considered appropriate for native compilation
without any intervention.

In other words, `SAVI_SYS_ROOT` is intended mostly for cross-compiling,
though it can also assist with selecting a specific SDK on MacOSX
where you may have multiple versions of the MacOSX SDK available.

The `SDK_ROOT` env var previously used for a similar purpose is now
deprecated, and will be fully removed in a future version of Savi.
But for now it is a synonym for the new `SAVI_SYS_ROOT` behavior.