Fix memory safety of `each`-style iteration methods.
Prior to this commit, these yielding methods were vulnerable
to memory unsafety issues if the length of the iterated collection
got reduced by something the caller did inside of the yield block.

This commit fixes these issues to make sure the internal loop
iterations are always bounded to the current `_size` of the collection.