Change Bytes and String methods to prefer `USize` over `ISize`.
The previous implementation of these methods copied from the Pony
standard library the convention of using `ISize` for specifying
a range in this kind of data structure, treating negative offsets
as being offsets from the end of the collection.

This change makes it so that they always use `USize`, making it
easier to have compatible values with other methods that use `USize`,
so as to avoid needing to translate between `USize` and `ISize`.

Now, if you want to do an offset from the end of the collection,
you'll need to do it yourself at the caller side.