Fix an LLVM IR issue with recursive yielding calls.
This use case (yielding functions that recursively call themselves)
was broken in the recent work done to upgrade Savi to LLVM 15 (#417),
which included fairly extensive changes to how pointers work in LLVM.

The breakage wasn't noticed until CI was run for the JSON library.

This PR fixes the issue and adds appropriate language-level tests
to prevent future regressions of this use case.