Cleanup: Remove temporary Time library patch.
This patch was temporarily introduced to make CI work for the
previous release, which included a breaking change that broke
the `Time` library, which is used in spec running for the core specs.

Now that the `Time` library library is up to date with the change,
we can remove this patch and go back to normal here.